# Conflicts:
#	src/main/java/com/atsuishio/superbwarfare/client/overlay/DroneHudOverlay.java
#	src/main/java/com/atsuishio/superbwarfare/client/overlay/JavelinHudOverlay.java
#	src/main/java/com/atsuishio/superbwarfare/config/server/MiscConfig.java
#	src/main/java/com/atsuishio/superbwarfare/config/server/VehicleConfig.java
#	src/main/java/com/atsuishio/superbwarfare/datagen/ModItemModelProvider.java
#	src/main/java/com/atsuishio/superbwarfare/entity/projectile/JavelinMissileEntity.java
#	src/main/java/com/atsuishio/superbwarfare/entity/projectile/RpgRocketTBGEntity.java
#	src/main/java/com/atsuishio/superbwarfare/init/ModItems.java
#	src/main/java/com/atsuishio/superbwarfare/item/common/ammo/RpgRocketStandard.java
#	src/main/java/com/atsuishio/superbwarfare/item/common/ammo/RpgRocketTBG.java
#	src/main/java/com/atsuishio/superbwarfare/item/gun/launcher/JavelinItem.java
#	src/main/resources/assets/superbwarfare/lang/en_us.json
#	src/main/resources/assets/superbwarfare/lang/ru_ru.json
#	src/main/resources/assets/superbwarfare/lang/uk_ua.json
#	src/main/resources/assets/superbwarfare/textures/entity/bmp2.png
#	src/main/resources/data/superbwarfare/recipes/compat/patchouli/superb_warfare_manual_crafting.json

+++ src/main/java/com/atsuishio/superbwarfare/entity/C4Entity.java
@@ -415,11 +415,11 @@ public class C4Entity extends Entity implements GeoEntity, OwnableEntity {
             }
         }
 
-        if (this.level() instanceof ServerLevel) {
+        if (this.level() instanceof ServerLevel && ExplosionConfig.EXPLOSION_DESTROY.get() && ExplosionConfig.EXTRA_EXPLOSION_EFFECT.get()) {
             AABB aabb = new AABB(pos, pos).inflate(2);
             BlockPos.betweenClosedStream(aabb).forEach((blockPos) -> {
                 float hard = this.level().getBlockState(blockPos).getBlock().defaultDestroyTime();
-                if (ExplosionConfig.EXPLOSION_DESTROY.get() && hard != -1) {
+                if (hard != -1) {
                     this.level().destroyBlock(blockPos, true);
                 }
             });
@@ -430,7 +430,6 @@ public class C4Entity extends Entity implements GeoEntity, OwnableEntity {
                 .damage(ExplosionConfig.C4_EXPLOSION_DAMAGE.get())
                 .radius(ExplosionConfig.C4_EXPLOSION_RADIUS.get())
                 .position(pos)
-                .causeVanillaExplosion()
                 .withParticleType(ParticleTool.ParticleType.HUGE)
                 .explode();
